Seperate convinience in ext and pref to avoid importing Gdk and Gtk in gnome-shell
