+++ display-brightness-ddcutil@themightydeity.github.com/conveniencePref.js
@@ -26,83 +26,7 @@
 */
 
 import Gdk from 'gi://Gdk';
-import Gio from 'gi://Gio';
-import GLib from 'gi://GLib';
 import Gtk from 'gi://Gtk';
-
-
-export function brightnessLog(settings, str) {
-    if (settings.get_boolean('verbose-debugging'))
-        log(`display-brightness-ddcutil extension:\n${str}`);
-}
-
-// timer
-/**
- * Taken from: https://github.com/optimisme/gjs-examples/blob/master/assets/timers.js
- */
-export function setTimeout(func, millis /* , ... args */) {
-    let args = [];
-    if (arguments.length > 2)
-        args = args.slice.call(arguments, 2);
-
-
-    const id = GLib.timeout_add(GLib.PRIORITY_DEFAULT, millis, () => {
-        func.apply(null, args);
-        return GLib.SOURCE_REMOVE; // Stop repeating
-    });
-
-    return id;
-}
-
-export function clearTimeout(id) {
-    GLib.source_remove(id);
-}
-
-
-export function setInterval(settings, func, millis /* , ... args */) {
-    let args = [];
-    if (arguments.length > 2)
-        args = args.slice.call(arguments, 2);
-
-    brightnessLog(settings, `setInterval called ${millis}`);
-
-    const id = GLib.timeout_add(GLib.PRIORITY_DEFAULT, millis, () => {
-        func.apply(null, args);
-        return GLib.SOURCE_CONTINUE; // Repeat
-    });
-
-    brightnessLog(settings, `setInterval set ${millis}`);
-
-    return id;
-}
-
-export function clearInterval(id) {
-    GLib.source_remove(id);
-}
-
-export function spawnWithCallback(settings, argv, callback) {
-    const proc = Gio.Subprocess.new(argv, Gio.SubprocessFlags.STDOUT_PIPE | Gio.SubprocessFlags.STDERR_SILENCE);
-
-    proc.communicate_utf8_async(null, null, (proc, res) => {
-        try {
-            const [, stdout, stderr] = proc.communicate_utf8_finish(res);
-            if (proc.get_successful()) {
-                callback(stdout);
-            } else {
-                /*
-                    errors from ddcutil (like monitor not found) were actually in stdout
-                    only the process return code was 1
-                */
-                if (stderr)
-                    callback(stderr);
-                else if (stdout)
-                    callback(stdout);
-            }
-        } catch (e) {
-            brightnessLog(settings, e.message);
-        }
-    });
-}
 
 /**
  * Check if the given keyval is forbidden.
